break up cartesian graph_interact.js:
- into component/fx/
- keep old 'init' routine in graph_interact.js
- mv hover constants out of cartesian constants.js and into fx/
