+++ src/plotly.js
@@ -24,8 +24,10 @@ exports.defaultConfig = require('./plot_api/plot_config');
 // plots
 exports.Plots = require('./plots/plots');
 exports.Axes = require('./plots/cartesian/axes');
-exports.Fx = require('./plots/cartesian/graph_interact');
+
+// components
 exports.ModeBar = require('./components/modebar');
+exports.Fx = require('./components/fx');
 
 // plot api
 require('./plot_api/plot_api');
